Add range and mods to YcmToggleLogs
Range allows a custom height to be specified.

Mods allows things like :vertical YcmToggleLogs. This is most useful
when used with a range, like :vertical 90YcmToggleLogs. This simply
allows to open the logs in a vertical split which is useful for large
logs.

You can also :botright etc. See :help <mods>.